requirements.txt
gunicorn>=21.2.0
flask>=3.0.0
azure-cosmos>=4.5.1
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
azure-ai-formrecognizer>=3.3.0
azure-core>=1.29.5
openai>=1.3.0
sentence-transformers>=2.7.0
redis>=5.0.1
gptcache>=0.1.43
PyMuPDF>=1.23.0
pytesseract>=0.3.10
pillow>=10.0.0
pdf2image>=1.16.3
python-dotenv>=1.0.1
requests>=2.31.0
opencensus-ext-azure>=1.1.0
azure-eventgrid